Bugfix 237: Warn on extra data in JSON stream instead of raising exception (#247)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] The commit message follows our guidelines:
https://github.com/liveloveapp/hashbrown/blob/main/CONTRIBUTING.md#commit
- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

```
[X] Bugfix
[ ] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Build related changes
[ ] CI related changes
[ ] Documentation content changes
[ ] Other... Please describe:
```

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->
Current versions of Hashbrown's streaming parser (aka Tater) will raise
an exception (out of the parser) when extra data is detected (meaning
Tater has finished a full object, but there are more characters in the
parsed string).

In practice, this occurs when the LLM returns a string with extra
objects, malformed objects or objects with unnecessary amounts of JSON
escaping, and this usually has to be resolved by prompt engineering
(with few-shot examples, etc). A user cannot fix the situation on their
own, and it will probably lead to incorrect behavior in the app.

Closes #237 

## What is the new behavior?
This PR removes the exception and adds a console.warn that describes the
funky situation and some ways to mitigate it.

To make sure the warn always shows up if the situation presents itself,
the warn does not go through the parser's logging mechanism.

To keep from warning on each chunk of a streaming message after extra
data starts showing up, a flag is added to the parser to indicate
"assume this is the final (complete) version of this string", and that
flag must be true for the warn to show up. In places in Hashbrown that
call the parser, the flag is false when dealing with a
currently-being-streamed message and true for "completed" messages.

## Does this PR introduce a breaking change?

```
[ ] Yes
[X] No
```

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information